+++ lib/api_request.dart
@@ -6,3 +6,4 @@ export 'src/api_request.dart';
 export 'src/api_request_client.dart';
 export 'src/api_request_error.dart';
 export 'src/api_request_options.dart';
+export 'src/interceptors/api_interceptor.dart';