+++ lib/src/api_request_options.dart
@@ -1,9 +1,13 @@
+import '../api_request.dart';
+
 typedef GetOption<T> = T Function();
 typedef GetAsyncOption<T> = Future<T> Function();
 
 class ApiRequestOptions {
-  static String bearer = 'Bearer ';
-  String? tokenType = '';
+  /// Singleton for ApiRequestOptions by create one instance
+  static ApiRequestOptions? _instance;
+
+  /// Singleton for ApiRequestOptions by create one instance
   static ApiRequestOptions? get instance {
     if (_instance == null) {
       _instance = ApiRequestOptions();
@@ -11,7 +15,22 @@ class ApiRequestOptions {
     return _instance;
   }
 
-  static ApiRequestOptions? _instance;
+  /// list of global interceptors
+  List<ApiInterceptor> interceptors = <ApiInterceptor>[];
+
+  /// to disable log set it to false
+  bool enableLog = true;
+
+  /// Timeout in milliseconds for opening url.
+  /// [Dio] will throw the [DioError] with [DioErrorType.connectTimeout] type
+  ///  when time out.
+  int? connectTimeout = 0;
+
+  /// for Bearer token type.
+  static String bearer = 'Bearer ';
+
+  /// set tokenType for types
+  String? tokenType = '';
 
   /// base url for your api
   /// You can set it ApiRequestOptions.instance.baseUrl = 'https://example.com';
@@ -29,6 +48,7 @@ class ApiRequestOptions {
   /// get access token use async callback function
   GetAsyncOption<String?>? getAsyncToken;
 
+  // set default query parameters to url
   late Map<String, dynamic> defaultQueryParameters = {};
 
   void config(
@@ -38,14 +58,31 @@ class ApiRequestOptions {
       GetAsyncOption<String?>? getAsyncToken,
       GetOption? unauthenticated,
       Map<String, dynamic>? defaultQueryParameters,
-      String? tokenType}) async {
+      String? tokenType,
+      int? connectTimeout,
+      bool? enableLog,
+      List<ApiInterceptor>? interceptors}) async {
     this.baseUrl = baseUrl ?? this.baseUrl;
     this.token = token ?? this.token;
     this.getToken = getToken ?? this.getToken;
     this.getAsyncToken = getAsyncToken ?? this.getAsyncToken;
     this.unauthenticated = unauthenticated ?? this.unauthenticated;
-    this.defaultQueryParameters =
-        defaultQueryParameters ?? this.defaultQueryParameters;
+
+    if (this.defaultQueryParameters.isNotEmpty) {
+      this.defaultQueryParameters.addAll(defaultQueryParameters ?? {});
+    } else {
+      this.defaultQueryParameters =
+          defaultQueryParameters ?? this.defaultQueryParameters;
+    }
+
+    if (this.interceptors.isNotEmpty) {
+      this.interceptors.addAll(interceptors ?? []);
+    } else {
+      this.interceptors = interceptors ?? this.interceptors;
+    }
+
     this.tokenType = tokenType ?? this.tokenType;
+    this.connectTimeout = connectTimeout ?? this.connectTimeout;
+    this.enableLog = enableLog ?? this.enableLog;
   }
 }