+++ lib/src/interceptors/api_interceptor.dart
@@ -1,0 +1,3 @@
+import 'package:dio/dio.dart';
+
+class ApiInterceptor extends Interceptor {}