+++ lib/src/interceptors/api_log_interceptor.dart
@@ -1,0 +1,128 @@
+import 'package:api_request/api_request.dart';
+import 'package:dio/dio.dart';
+
+class ApiLogInterceptor extends ApiInterceptor {
+  final String _tag = 'log';
+
+  ApiLogInterceptor({
+    this.request = true,
+    this.requestHeader = true,
+    this.requestBody = true,
+    this.responseHeader = true,
+    this.responseBody = true,
+    this.error = true,
+    this.logPrint = print,
+  });
+
+  /// Print request [Options]
+  bool request;
+
+  /// Print request header [Options.headers]
+  bool requestHeader;
+
+  /// Print request data [Options.data]
+  bool requestBody;
+
+  /// Print [Response.data]
+  bool responseBody;
+
+  /// Print [Response.headers]
+  bool responseHeader;
+
+  /// Print error message
+  bool error;
+
+  /// Log printer; defaults print log to console.
+  /// In flutter, you'd better use debugPrint.
+  /// you can also write log in a file, for example:
+  ///```dart
+  ///  var file=File("./log.txt");
+  ///  var sink=file.openWrite();
+  ///  dio.interceptors.add(LogInterceptor(logPrint: sink.writeln));
+  ///  ...
+  ///  await sink.close();
+  ///```
+  void Function(Object object) logPrint;
+
+  @override
+  void onRequest(
+      RequestOptions options, RequestInterceptorHandler handler) async {
+    logPrint('*** Api Request ***');
+    _printKV('uri', options.uri);
+
+    if (request) {
+      _printKV('method', options.method);
+      _printKV('responseType', options.responseType.toString());
+      _printKV('connectTimeout', options.connectTimeout);
+    }
+    if (requestHeader) {
+      logPrint('headers:');
+      options.headers.forEach((key, v) => _printKV(' $key', v));
+    }
+    if (requestBody) {
+      logPrint('data:');
+      _printAll(options.data);
+    }
+    logPrint('');
+
+    handler.next(options);
+  }
+
+  @override
+  void onResponse(Response response, ResponseInterceptorHandler handler) async {
+    logPrint('*** Api Response ***');
+    _printResponse(response);
+    handler.next(response);
+  }
+
+  @override
+  void onError(DioError err, ErrorInterceptorHandler handler) async {
+    if (error) {
+      logPrint('*** ApiRequestError ***:');
+      logPrint('uri: ${err.requestOptions.uri}');
+      logPrint('$err');
+      if (err.response != null) {
+        _printResponse(err.response!);
+      }
+      logPrint('');
+    }
+
+    handler.next(err);
+  }
+
+  void _printResponse(Response response) {
+    _printKV('uri', response.requestOptions.uri);
+    if (responseHeader) {
+      _printKV('statusCode', response.statusCode);
+      if (response.isRedirect == true) {
+        _printKV('redirect', response.realUri);
+      }
+
+      logPrint('headers:');
+      response.headers.forEach((key, v) => _printKV(' $key', v.join('\r\n\t')));
+    }
+    if (responseBody) {
+      logPrint('Response Text:');
+      _printAll(response.toString());
+    }
+    logPrint('');
+  }
+
+  void _printKV(String key, Object? v) {
+    logPrint('$key: $v');
+  }
+
+  void _printAll(msg) {
+    msg.toString().split('\n').forEach(logPrint);
+  }
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+      other is ApiLogInterceptor &&
+          runtimeType == other.runtimeType &&
+          _tag == other._tag;
+
+  @override
+  int get hashCode => _tag.hashCode;
+}